# docs/智能测试平台.openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "智能测试平台 API",
    "description": "智能测试平台完整接口文档，包含认证、项目管理、需求管理、测试点管理、测试用例管理、自动化执行六大模块",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "本地开发服务器"
    }
  ],
  "paths": {
...
    "/api/requirements": {
      "get": {
        "summary": "获取所有需求",
        "description": "获取所有需求列表",
        "operationId": "getAllRequirements",
        "tags": ["需求管理模块"],
        "security": [{"BasicAuth": []}],
        "responses": {
          "200": {
            "description": "成功获取需求列表",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Requirement"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/requirements/project/{projectId}": {
      "get": {
        "summary": "按项目获取需求",
        "description": "获取指定项目下的所有需求",
        "operationId": "getRequirementsByProjectId",
        "tags": ["需求管理模块"],
        "security": [{"BasicAuth": []}],
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "description": "项目ID"
          }
        ],
        "responses": {
          "200": {
            "description": "成功获取需求列表",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Requirement"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/requirements/{id}": {
      "get": {
        "summary": "获取需求详情",
        "description": "根据ID获取需求详细信息",
        "operationId": "getRequirementById",
        "tags": ["需求管理模块"],
        "security": [{"BasicAuth": []}],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "description": "需求ID"
          }
        ],
        "responses": {
          "200": {
            "description": "成功获取需求详情",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Requirement"
                }
              }
            }
          }
        }
      }
    },
    "/api/requirements/upload": {
      "post": {
        "summary": "上传需求文件",
        "description": "上传需求文档文件",
        "operationId": "uploadRequirement",
        "tags": ["需求管理模块"],
        "security": [{"BasicAuth": []}],
        "parameters": [
          {
            "name": "projectId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "description": "项目ID"
          },
          {
            "name": "name",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "需求名称"
          },
          {
            "name": "version",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "版本号"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "description": "需求文件"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "上传成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Requirement"
                }
              }
            }
          }
        }
      }
    },
    "/api/requirements/{id}/generate-test-points": {
      "post": {
        "summary": "生成测试点",
        "description": "根据需求自动生成测试点（AI分析）",
        "operationId": "generateTestPoints",
        "tags": ["需求管理模块"],
        "security": [{"BasicAuth": []}],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "description": "需求ID"
          }
        ],
        "responses": {
          "200": {
            "description": "生成成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TestPoint"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/requirements/{id}/confirm-test-points": {
      "post": {
        "summary": "确认测试点",
        "description": "确认需求关联的测试点",
        "operationId": "confirmTestPoints",
        "tags": ["需求管理模块"],
        "security": [{"BasicAuth": []}],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "description": "需求ID"
          }
        ],
        "responses": {
          "200": {
            "description": "确认成功"
          }
        }
      }
    },
...
    "/api/test-points/requirement/{requirementId}": {
      "get": {
        "summary": "按需求获取测试点",
        "description": "获取指定需求下的所有测试点",
        "operationId": "getTestPointsByRequirementId",
        "tags": ["测试点管理模块"],
        "security": [{"BasicAuth": []}],
        "parameters": [
          {
            "name": "requirementId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "description": "需求ID"
          }
        ],
        "responses": {
          "200": {
            "description": "成功获取测试点列表",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TestPoint"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/test-cases": {
      "get": {
        "summary": "获取所有测试用例",
        "description": "获取所有测试用例列表",
        "operationId": "getAllTestCases",
        "tags": ["测试用例管理模块"],
        "security": [{"BasicAuth": []}],
        "responses": {
          "200": {
            "description": "成功获取测试用例列表",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TestCase"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "创建测试用例",
        "description": "手动创建测试用例",
        "operationId": "addTestCase",
        "tags": ["测试用例管理模块"],
        "security": [{"BasicAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTestCaseRequest"
              },
              "example": {
                "projectId": 1,
                "testPointId": 1,
                "title": "登录成功用例",
                "precondition": "用户已注册",
                "steps": "1. 打开登录页面\n2. 输入正确的用户名和密码\n3. 点击登录按钮",
                "expectedResult": "登录成功，跳转到首页",
                "priority": "HIGH",
                "caseType": "FUNCTIONAL",
                "automation": true
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "创建成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TestCase"
                }
              }
            }
          }
        }
      }
    },
    "/api/test-cases/project/{projectId}": {
      "get": {
        "summary": "按项目获取测试用例",
        "description": "获取指定项目下的所有测试用例",
        "operationId": "getTestCasesByProjectId",
        "tags": ["测试用例管理模块"],
        "security": [{"BasicAuth": []}],
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "description": "项目ID"
          }
        ],
        "responses": {
          "200": {
            "description": "成功获取测试用例列表",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TestCase"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/test-cases/test-point/{testPointId}": {
      "get": {
        "summary": "按测试点获取测试用例",
        "description": "获取指定测试点下的所有测试用例",
        "operationId": "getTestCasesByTestPointId",
        "tags": ["测试用例管理模块"],
        "security": [{"BasicAuth": []}],
        "parameters": [
          {
            "name": "testPointId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "description": "测试点ID"
          }
        ],
        "responses": {
          "200": {
            "description": "成功获取测试用例列表",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TestCase"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/test-cases/{id}": {
      "get": {
        "summary": "获取测试用例详情",
        "description": "根据ID获取测试用例详细信息",
        "operationId": "getTestCaseById",
        "tags": ["测试用例管理模块"],
        "security": [{"BasicAuth": []}],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "description": "用例ID"
          }
        ],
        "responses": {
          "200": {
            "description": "成功获取测试用例详情",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TestCase"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "更新测试用例",
        "description": "更新测试用例信息",
        "operationId": "updateTestCase",
        "tags": ["测试用例管理模块"],
        "security": [{"BasicAuth": []}],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "description": "用例ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTestCaseRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "更新成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TestCase"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "删除测试用例",
        "description": "删除指定测试用例",
        "operationId": "deleteTestCase",
        "tags": ["测试用例管理模块"],
        "security": [{"BasicAuth": []}],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "description": "用例ID"
          }
        ],
        "responses": {
          "200": {
            "description": "删除成功"
          }
        }
      }
    },
    "/api/test-cases/generate": {
      "post": {
        "summary": "生成测试用例（单个）",
        "description": "根据测试点自动生成测试用例（AI生成）",
        "operationId": "generateTestCases",
        "tags": ["测试用例管理模块"],
        "security": [{"BasicAuth": []}],
        "parameters": [
          {
            "name": "projectId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "description": "项目ID"
          },
          {
            "name": "testPointId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "description": "测试点ID"
          }
        ],
        "responses": {
          "200": {
            "description": "生成成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TestCase"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/test-cases/generate/batch": {
      "post": {
        "summary": "批量生成测试用例",
        "description": "根据多个测试点批量生成测试用例（AI生成）",
        "operationId": "generateBatchTestCases",
        "tags": ["测试用例管理模块"],
        "security": [{"BasicAuth": []}],
        "parameters": [
          {
            "name": "projectId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "description": "项目ID"
          },
          {
            "name": "testPointIds",
            "in": "query",
            "required": true,
            "schema": {
              "type": "array",
              "items": {
                "type": "integer",
                "format": "int64"
              }
            },
            "description": "测试点ID列表"
          }
        ],
        "responses": {
          "200": {
            "description": "生成成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TestCase"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Requirement": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "description": "需求ID"
          },
          "projectId": {
            "type": "integer",
            "format": "int64",
            "description": "所属项目ID"
          },
          "name": {
            "type": "string",
            "description": "需求名称"
          },
          "fileUrl": {
            "type": "string",
            "description": "需求文件URL"
          },
          "version": {
            "type": "string",
            "description": "版本号"
          },
          "status": {
            "type": "string",
            "enum": ["PENDING_ANALYSIS", "ANALYZING", "ANALYZED", "CONFIRMED"],
            "description": "状态"
          },
          "testPointCount": {
            "type": "integer",
            "format": "int32",
            "description": "测试点数量"
          },
          "owner": {
            "type": "string",
            "description": "负责人"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "创建时间"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "更新时间"
          }
        }
      },
      "TestPoint": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "description": "测试点ID"
          },
          "projectId": {
            "type": "integer",
            "format": "int64",
            "description": "所属项目ID"
          },
          "requirementId": {
            "type": "integer",
            "format": "int64",
            "description": "所属需求ID"
          },
          "requirementItemId": {
            "type": "integer",
            "format": "int64",
            "description": "需求项ID"
          },
          "pointCode": {
            "type": "string",
            "description": "测试点编码"
          },
          "pointName": {
            "type": "string",
            "description": "测试点名称"
          },
          "pointType": {
            "type": "string",
            "description": "测试点类型"
          },
          "name": {
            "type": "string",
            "description": "名称"
          },
          "function": {
            "type": "string",
            "description": "功能描述"
          },
          "sceneType": {
            "type": "string",
            "description": "场景类型"
          },
          "riskLevel": {
            "type": "string",
            "description": "风险等级"
          },
          "businessRule": {
            "type": "string",
            "description": "业务规则"
          },
          "description": {
            "type": "string",
            "description": "描述"
          },
          "automationSuggested": {
            "type": "boolean",
            "description": "是否建议自动化"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "创建时间"
          },
          "createdBy": {
            "type": "integer",
            "format": "int64",
            "description": "创建人ID"
          }
        }
      },
...
      "TestCase": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "description": "用例ID"
          },
          "projectId": {
            "type": "integer",
            "format": "int64",
            "description": "所属项目ID"
          },
          "testPointId": {
            "type": "integer",
            "format": "int64",
            "description": "所属测试点ID"
          },
          "caseType": {
            "type": "string",
            "enum": ["FUNCTIONAL", "API", "UI", "PERFORMANCE"],
            "description": "用例类型"
          },
          "caseNumber": {
            "type": "string",
            "description": "用例编号"
          },
          "caseCode": {
            "type": "string",
            "description": "用例编码"
          },
          "title": {
            "type": "string",
            "description": "用例标题"
          },
          "precondition": {
            "type": "string",
            "description": "前置条件"
          },
          "steps": {
            "type": "string",
            "description": "测试步骤"
          },
          "expectedResult": {
            "type": "string",
            "description": "预期结果"
          },
          "priority": {
            "type": "string",
            "enum": ["HIGH", "MEDIUM", "LOW"],
            "description": "优先级"
          },
          "automation": {
            "type": "boolean",
            "description": "是否自动化"
          },
          "status": {
            "type": "string",
            "enum": ["PENDING_REVIEW", "REVIEWED", "APPROVED", "REJECTED"],
            "description": "状态"
          },
          "creator": {
            "type": "string",
            "description": "创建人"
          },
          "createdBy": {
            "type": "integer",
            "format": "int64",
            "description": "创建人ID"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "创建时间"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "更新时间"
          }
        }
      },
      "CreateTestCaseRequest": {
        "type": "object",
        "required": ["projectId", "testPointId", "title"],
        "properties": {
          "projectId": {
            "type": "integer",
            "format": "int64",
            "description": "所属项目ID"
          },
          "testPointId": {
            "type": "integer",
            "format": "int64",
            "description": "所属测试点ID"
          },
          "title": {
            "type": "string",
            "description": "用例标题"
          },
          "precondition": {
            "type": "string",
            "description": "前置条件"
          },
          "steps": {
            "type": "string",
            "description": "测试步骤"
          },
          "expectedResult": {
            "type": "string",
            "description": "预期结果"
          },
          "priority": {
            "type": "string",
            "description": "优先级"
          },
          "caseType": {
            "type": "string",
            "description": "用例类型"
          },
          "automation": {
            "type": "boolean",
            "description": "是否自动化"
          }
        }
      },
...
    },
    "securitySchemes": {
      "BasicAuth": {
        "type": "http",
        "scheme": "basic",
        "description": "Basic认证，格式为 `Basic Base64(username:password)`"
      }
    }
  },
  "security": [
    {
      "BasicAuth": []
    }
  ]
}